pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "qrcode>=8.1",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
]
